Properties are now called whenever Initialization of ENT is called.
Some entities may change their properties during their Initialization.
It is best to change properties such as skin after spawning an entity.
Imagine an entity that sets skin during Initialization (Or changes position or anges).
Then, the modification of user will be completely ignored if properties are changed first and then the prop is spawned.